perbaikan dashboard

+++ app/Http/Controllers/DashboardController.php
@@ -44,7 +44,7 @@ class DashboardController extends Controller
             $nilai = [];
             foreach($kitabs as $k){
                 $kitab[] = $k->title;
-                $nilai[] = var_dump((float) Nilai::where('kitab_id', $k->id)->avg('nilai'));
+                $nilai[] = var_dump(Nilai::where('kitab_id', $k->id)->avg('nilai'));
             }
 
             dd($nilai);
